que bvien arregle muchas cosas

+++ database/migrations/2024_10_18_162041_create_services_table.php
@@ -18,9 +18,9 @@ return new class extends Migration
             $table->string('name', 50);
             $table->string('description', 255);
             $table->foreignId('service_category_id')->constrained('service_categories');
-            $table->decimal('price', 10, 2);
+            $table->integer('price');
             $table->integer('people_quantity')->nullable();
-            $table->decimal('coast', 10, 2)->nullable();
+            $table->integer('coast')->nullable();
             $table->string('image_path')->default('https://res.cloudinary.com/ddclkt7n4/image/upload/v1733793457/unpsg410bmgyadi8edr8.jpg')->nullable();
             $table->boolean('is_active')->default(true);
             $table->boolean('quantifiable')->default(false);
